4-gallon bucket problem
The "4 gallon bucket problem" often refers to a classic type of puzzle known as a water jug problem, which has been popularized in movies and puzzle books. Here’s a common version of the problem:

You are given two buckets, one with a capacity of 3 gallons and another with a capacity of 4 gallons, and you must use these buckets to measure exactly 2 gallons of water. You can fill up a bucket, empty a bucket, and pour water from one bucket to another either until one of the buckets is full or the other is empty.

The challenge lies in figuring out a sequence of actions that results in exactly 2 gallons of water in one of the buckets. This puzzle is a test of logical reasoning and has been used in various forms to teach problem-solving techniques.